+++ src/index.mjs
@@ -47,4 +47,4 @@ export { BasisManager } from './extras/BasisManager.js';
 export { AxesHelper } from './extras/helpers/AxesHelper.js';
 export { GridHelper } from './extras/helpers/GridHelper.js';
 export { VertexNormalsHelper } from './extras/helpers/VertexNormalsHelper.js';
-export { FacesNormalsHelper } from './extras/helpers/FacesNormalsHelper.js';
+export { FaceNormalsHelper } from './extras/helpers/FaceNormalsHelper.js';
